Don't test symlinks on windows for now
we are currently unable to resolve them, since we rely on os.path.realpath for that and it doesn't resolve windows' symlinks

+++ datalad/distribution/tests/test_dataset.py
@@ -379,46 +379,51 @@ def test_property_reevaluation(repo1, repo2, repo3, non_repo, symlink):
     assert (ds.id != first_id)
     second_id = ds.id
 
-    # now, let ds be a symlink and change that symlink to point to different
-    # things:
-    ar2 = AnnexRepo(repo2)
-    ar3 = AnnexRepo(repo3)
-    assert (os.path.isabs(non_repo))
+    # While os.symlink does work on windows (since vista), os.path.realpath
+    # doesn't resolve such symlinks. This has all kinds of implications.
+    # Hopefully this can be dealt with, when we switch to using pathlib
+    # (see datalad-revolution).
+    if not on_windows:
+        # now, let ds be a symlink and change that symlink to point to different
+        # things:
+        ar2 = AnnexRepo(repo2)
+        ar3 = AnnexRepo(repo3)
+        assert (os.path.isabs(non_repo))
 
-    os.symlink(repo1, symlink)
-    ds_link = Dataset(symlink)
-    assert (ds_link.repo is ds.repo)  # same Repo instance
-    assert (ds_link is not ds)  # but not the same Dataset instance
-    assert (ds_link.config is ds.repo.config)
-    assert (ds_link._cfg_bound is True)
-    assert (ds_link.id is not None)
-    # same id, although different Dataset instance:
-    assert (ds_link.id == second_id)
+        os.symlink(repo1, symlink)
+        ds_link = Dataset(symlink)
+        assert (ds_link.repo is ds.repo)  # same Repo instance
+        assert (ds_link is not ds)  # but not the same Dataset instance
+        assert (ds_link.config is ds.repo.config)
+        assert (ds_link._cfg_bound is True)
+        assert (ds_link.id is not None)
+        # same id, although different Dataset instance:
+        assert (ds_link.id == second_id)
 
-    os.unlink(symlink)
-    os.symlink(repo2, symlink)
+        os.unlink(symlink)
+        os.symlink(repo2, symlink)
 
-    assert (ds_link.repo is ar2)  # same Repo instance
-    assert (ds_link.config is ar2.config)
-    assert (ds_link._cfg_bound is True)
-    # id is None again, since this repository is an annex but there was no
-    # Dataset.create() called yet.
-    assert (ds_link.id is None)
+        assert (ds_link.repo is ar2)  # same Repo instance
+        assert (ds_link.config is ar2.config)
+        assert (ds_link._cfg_bound is True)
+        # id is None again, since this repository is an annex but there was no
+        # Dataset.create() called yet.
+        assert (ds_link.id is None)
 
-    os.unlink(symlink)
-    os.symlink(repo3, symlink)
+        os.unlink(symlink)
+        os.symlink(repo3, symlink)
 
-    assert (ds_link.repo is ar3)  # same Repo instance
-    assert (ds_link.config is ar3.config)
-    assert (ds_link._cfg_bound is True)
-    # id is None again, since this repository is an annex but there was no
-    # Dataset.create() called yet.
-    assert (ds_link.id is None)
+        assert (ds_link.repo is ar3)  # same Repo instance
+        assert (ds_link.config is ar3.config)
+        assert (ds_link._cfg_bound is True)
+        # id is None again, since this repository is an annex but there was no
+        # Dataset.create() called yet.
+        assert (ds_link.id is None)
 
-    os.unlink(symlink)
-    os.symlink(non_repo, symlink)
+        os.unlink(symlink)
+        os.symlink(non_repo, symlink)
 
-    assert (ds_link.repo is None)
-    assert (ds_link.config is not ar3.config)
-    assert (ds_link._cfg_bound is False)
-    assert (ds_link.id is None)
+        assert (ds_link.repo is None)
+        assert (ds_link.config is not ar3.config)
+        assert (ds_link._cfg_bound is False)
+        assert (ds_link.id is None)
